refactor server impl
