+++ pkg/server/user_test.go
@@ -6,27 +6,27 @@ import (
 )
 
 type UserTestSuite struct {
-    ServerTestSuite
+	ServerTestSuite
 }
 
 //func (s *UserTestSuite) TestGetUser() {
-//	req := v1.GetUserRequest{}
-//	resp, err := s.testServer.GetUser(context.Background(), &req)
-//	if s.NoError(err) {
-//		expectedUserProfile := expectedUserProfiles[0]
-//		s.Equal(expectedUserProfile.User.PreferredOrganization, resp.OrganizationId)
-//		s.Equal(expectedUserProfile.User.ID, resp.Id)
-//		s.Equal(fmt.Sprintf("%s %s",
-//			expectedUserProfile.User.FirstName, expectedUserProfile.User.LastName), resp.Name)
-//		s.Contains(s.mockPennsieve.JWTToAPIKey, resp.SessionToken)
-//		s.Equal(expectedUserProfile.Profile.APIToken, s.mockPennsieve.JWTToAPIKey[resp.SessionToken])
+//req := v1.GetUserRequest{}
+//resp, err := s.testServer.GetUser(context.Background(), &req)
+//if s.NoError(err) {
+//	expectedUserProfile := expectedUserProfiles[0]
+//	s.Equal(expectedUserProfile.User.PreferredOrganization, resp.OrganizationId)
+//	s.Equal(expectedUserProfile.User.ID, resp.Id)
+//	s.Equal(fmt.Sprintf("%s %s",
+//		expectedUserProfile.User.FirstName, expectedUserProfile.User.LastName), resp.Name)
+//	s.Contains(s.mockPennsieve.JWTToAPIKey, resp.SessionToken)
+//	s.Equal(expectedUserProfile.Profile.APIToken, s.mockPennsieve.JWTToAPIKey[resp.SessionToken])
 //
-//		s.Equal(expectedUserProfile.Profile.Name, resp.Profile)
-//		s.Equal(expectedUserProfile.Org.ID, resp.OrganizationId)
-//		s.Equal(expectedUserProfile.Org.Name, resp.OrganizationName)
-//		s.Equal(s.mockPennsieve.API.Server.URL, resp.ApiHost)
-//		s.NotEmpty(resp.Api2Host)
-//	}
+//	s.Equal(expectedUserProfile.Profile.Name, resp.Profile)
+//	s.Equal(expectedUserProfile.Org.ID, resp.OrganizationId)
+//	s.Equal(expectedUserProfile.Org.Name, resp.OrganizationName)
+//	s.Equal(s.mockPennsieve.API.Server.URL, resp.ApiHost)
+//	s.NotEmpty(resp.Api2Host)
+//}
 //}
 
 //
@@ -42,5 +42,5 @@ type UserTestSuite struct {
 //}
 
 func TestUserSuite(t *testing.T) {
-    suite.Run(t, new(UserTestSuite))
+	suite.Run(t, new(UserTestSuite))
 }
